apps/ai-service/requirements.txt
fastapi
uvicorn
faster-whisper
spacy
transformers
sentencepiece
sacremoses
torchaudio
python-multipart
pydantic
structlog
